backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "islom-sattorov",email = "slmsttrv1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "load-dotenv (>=0.1.0,<0.2.0)",
    "fastapi (>=0.121.0,<0.122.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.3,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-height = 100
